# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
export interface AcceptInviteBody {
  token: string;
}

export interface UserCompact {
  id: number;
  name: string;
  avatarUrl?: string | null;
}

export interface ActivityLogResponseItem {
  id: number;
  action: string;
  entityType: string;
  entityId: number;
  userId: number;
  workspaceId?: number | null;
  description?: string | null;
  user?: UserCompact | null;
  createdAt: string;
}

export interface TaskCountByLabel {
  label: string;
  count: number;
}

export interface AdminStatsResponse {
  totalUsers: number;
  totalWorkspaces: number;
  totalTasks: number;
  completedTasks: number;
  activeUsers: number;
  usersByRole?: TaskCountByLabel[] | null;
  tasksByStatus?: TaskCountByLabel[] | null;
}

export interface ChangePasswordBody {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface CreateProjectBody {
  /** @minLength 1 */
  name: string;
  description?: string | null;
  workspaceId: number;
}

export type TaskStatusEnum = typeof TaskStatusEnum[keyof typeof TaskStatusEnum];


export const TaskStatusEnum = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskPriorityEnum = typeof TaskPriorityEnum[keyof typeof TaskPriorityEnum];


export const TaskPriorityEnum = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface CreateTaskBody {
  /** @minLength 1 */
  title: string;
  description?: string | null;
  status?: TaskStatusEnum;
  priority?: TaskPriorityEnum;
  dueDate?: string | null;
  projectId: number;
  assigneeId?: number | null;
  labels?: string[] | null;
}

export interface CreateWorkspaceBody {
  /** @minLength 1 */
  name: string;
  description?: string | null;
}

export interface DashboardSummaryResponse {
  totalWorkspaces: number;
  totalTasks: number;
  myTasks: number;
  myPendingTasks: number;
  myCompletedTasks: number;
  recentActivity: ActivityLogResponseItem[];
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ValidationError {
  loc: (string | number)[];
  msg: string;
  type: string;
}

export interface HTTPValidationError {
  detail?: ValidationError[];
}

export type WorkspaceRoleEnum = typeof WorkspaceRoleEnum[keyof typeof WorkspaceRoleEnum];


export const WorkspaceRoleEnum = {
  owner: 'owner',
  admin: 'admin',
  member: 'member',
  viewer: 'viewer',
} as const;

export interface InviteMemberBody {
  email: string;
  role?: WorkspaceRoleEnum;
}

export type ProjectRoleEnum = typeof ProjectRoleEnum[keyof typeof ProjectRoleEnum];


export const ProjectRoleEnum = {
  admin: 'admin',
  member: 'member',
} as const;

export interface InviteProjectMemberBody {
  email: string;
  role?: ProjectRoleEnum;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserResponse {
  id: number;
  name: string;
  email: string;
  isSuperAdmin: boolean;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface LoginResponse {
  user: UserResponse;
  accessToken: string;
  refreshToken: string;
}

export interface MessageResponse {
  message: string;
}

export type NotificationTypeEnum = typeof NotificationTypeEnum[keyof typeof NotificationTypeEnum];


export const NotificationTypeEnum = {
  task_assigned: 'task_assigned',
  due_date_reminder: 'due_date_reminder',
  task_completed: 'task_completed',
  workspace_invite: 'workspace_invite',
  member_joined: 'member_joined',
} as const;

export interface NotificationResponseItem {
  id: number;
  type: NotificationTypeEnum;
  title: string;
  message: string;
  isRead: boolean;
  userId: number;
  taskId?: number | null;
  workspaceId?: number | null;
  createdAt: string;
}

export interface ProjectMemberResponse {
  userId: number;
  projectId: number;
  role: ProjectRoleEnum;
  joinedAt: string;
  user: UserResponse;
}

export interface ProjectResponseItem {
  id: number;
  name: string;
  description?: string | null;
  workspaceId: number;
  createdById: number;
  createdAt: string;
  updatedAt?: string | null;
  workspaceCount?: number | null;
  memberCount?: number | null;
  myRole?: ProjectRoleEnum | null;
}

export interface ResetPasswordBody {
  token: string;
  /**
     * Password must be at least 8 characters
     * @minLength 8
     */
  password: string;
}

export interface SignupBody {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface TaskCountByUser {
  userId: number;
  name: string;
  avatarUrl?: string | null;
  totalTasks: number;
  completedTasks: number;
}

export interface TaskResponseItem {
  id: number;
  title: string;
  description?: string | null;
  status: TaskStatusEnum;
  priority: TaskPriorityEnum;
  dueDate?: string | null;
  projectId: number;
  assigneeId?: number | null;
  createdById: number;
  labels?: string[] | null;
  assignee?: UserCompact | null;
  createdBy?: UserCompact | null;
  createdAt: string;
  updatedAt?: string | null;
}

export interface UpdateProfileBody {
  name?: string | null;
  avatarUrl?: string | null;
}

export interface UpdateProjectBody {
  name?: string | null;
  description?: string | null;
}

export interface UpdateProjectMemberRoleBody {
  role: ProjectRoleEnum;
}

export interface UpdateRoleBody {
  is_super_admin: boolean;
}

export interface UpdateTaskBody {
  title?: string | null;
  description?: string | null;
  status?: TaskStatusEnum | null;
  priority?: TaskPriorityEnum | null;
  dueDate?: string | null;
  assigneeId?: number | null;
  labels?: string[] | null;
}

export interface UpdateWorkspaceBody {
  name?: string | null;
  description?: string | null;
}

export interface WorkspaceAnalyticsResponse {
  totalTasks: number;
  completedTasks: number;
  pendingTasks: number;
  inProgressTasks: number;
  memberCount: number;
  completionRate: number;
  tasksByPriority: TaskCountByLabel[];
  tasksByMember: TaskCountByUser[];
}

export interface WorkspaceInvitationResponse {
  id: number;
  workspaceId: number;
  email: string;
  role: WorkspaceRoleEnum;
  token: string;
  expiresAt: string;
  accepted: boolean;
  createdAt: string;
}

export interface WorkspaceMemberResponse {
  userId: number;
  workspaceId: number;
  role: WorkspaceRoleEnum;
  joinedAt: string;
  user: UserResponse;
}

export interface WorkspaceResponseItem {
  id: number;
  name: string;
  description?: string | null;
  ownerId: number;
  memberCount?: number | null;
  taskCount?: number | null;
  createdAt: string;
}

export type UpdateMemberRoleParams = {
role: string;
};

export type SearchUsersParams = {
/**
 * Scope search to a specific workspace
 */
workspace_id: number;
/**
 * @minLength 0
 */
query?: string;
};